Abstract iframe registration to webcontents
