+++ src/browser/api/window.js
@@ -443,38 +443,6 @@ Window.create = function(id, opts) {
     if (!opts._noregister) {
 
         browserWindow = BrowserWindow.fromId(id);
-
-        // called in the WebContents class in the runtime
-        browserWindow.webContents.registerIframe = (frameName, frameRoutingId) => {
-            // called for all iframes, but not for main frame of windows
-            electronApp.vlog(1, `registerIframe ${frameName} ${frameRoutingId}`);
-            const parentFrameId = id;
-            const frameInfo = {
-                name: frameName,
-                uuid,
-                parentFrameId,
-                parent: { uuid, name },
-                frameRoutingId,
-                entityType: 'iframe'
-            };
-
-            winObj.frames.set(frameName, frameInfo);
-        };
-
-        // called in the WebContents class in the runtime
-        browserWindow.webContents.unregisterIframe = (closedFrameName, frameRoutingId) => {
-            // called for all iframes AND for main frames
-            electronApp.vlog(1, `unregisterIframe ${frameRoutingId} ${closedFrameName}`);
-            const frameName = closedFrameName || name; // the parent name is considered a frame as well
-            const frameInfo = winObj.frames.get(closedFrameName);
-            const entityType = frameInfo ? 'iframe' : 'window';
-            const payload = { uuid, name, frameName, entityType };
-
-            winObj.frames.delete(closedFrameName);
-            ofEvents.emit(route.frame('disconnected', uuid, closedFrameName), payload);
-            ofEvents.emit(route.window('frame-disconnected', uuid, name), payload);
-        };
-
         webContents = browserWindow.webContents;
 
         //Legacy 5.0 feature, if customWindowAlert flag is found all alerts will be suppresed,
@@ -1002,6 +970,7 @@ Window.create = function(id, opts) {
             name
         });
     }
+    WebContents.setIframeHandlers(browserWindow.webContents, winObj, uuid, name);
 
     return winObj;
 };
